kernel flags: make repartitioning and graph coloring optional
We introduce a small set of flags in phist_kernel_flags.h, and
each subroutine states which of them it potentially uses.
The decision wether to respect the flag is up to the kernel lib,
however.

One of the advantages is that we can avoid passing non-symmetric matrices
to Scotch in the tests.