(cmake) fixed optional Trilinos libs (Kokkos and Teuchos were missing)
Technically speaking, if Belos is compiled to use TSQR and used as an optional lib,
we should have Kokkos and Teuchos as requirements for phist+ghost. Now they are treated as optional,
however, I can't imagine a setup where Belos is found and Teuchos and Kokkos aren't, so it should work in practice.