Workaround for the missing phist_maps_compatible function with GHOST. The problem here is that GHOST stores permutations in the map, but the info wether a permutation has been applied to a vector/densemat in the densemat itself. So purely based on the maps we can't say if a vector has to be permuted. I think this calls for a re-thinking of the phist kernel interface regarding permutations. The mvec_to_mvec function, on the oter hand, should be quite easy to implement with GHOST 1.1 but hasn't been updated.
Re #196: it looks like we re-established compatibility with GHOST (1.1) except for the
functions phist_maps_compatible (always returns 0 right now) and mvec_to_mvec. Also, the CARP-CG
has to be looked at because it requires local permutations of the the col_map and I'm not sure if this
works already or if some tests are disabled.